Implemented game logic for a player move or undo
Added new methods into the GameLogic class for implementing a player move and an undo which updates the mancala after being performed. Updated the Mancala Frame to add action listeners to the pit buttons that call the playerMoved method when pressed as well as to the undo button that calls the undoMove method when pressed.